Fix/post: Add user Post Data (#35)

+++ src/ChitChat.Application/Services/PostService.cs
@@ -5,11 +5,9 @@ using AutoMapper;
 using ChitChat.Application.Exceptions;
 using ChitChat.Application.Helpers;
 using ChitChat.Application.Localization;
-using ChitChat.Application.Models;
 using ChitChat.Application.Models.Dtos.Post;
 using ChitChat.Application.Models.Dtos.Post.Comments;
 using ChitChat.Application.Models.Dtos.Post.CreatePost;
-using ChitChat.Application.Services.CloudinaryInterface;
 using ChitChat.Application.Services.Interface;
 using ChitChat.DataAccess.Repositories.Interface;
 using ChitChat.DataAccess.Repositories.Interrface;
@@ -17,7 +15,6 @@ using ChitChat.Domain.Entities.PostEntities;
 using ChitChat.Domain.Entities.PostEntities.Reaction;
 using ChitChat.Domain.Entities.SystemEntities;
 using ChitChat.Domain.Enums;
-using ChitChat.Domain.Extensions;
 
 using Microsoft.EntityFrameworkCore;
 
@@ -54,48 +51,82 @@ namespace ChitChat.Application.Services
             _cloudinaryService = cloudinaryService;
         }
         #region GET
-        public async Task<List<PostDto>> GetAllPostsAsync(PostSearchQueryDto queryDto)
+        public async Task<List<PostDto>> GetAllPostsAsync(PostUserSearchQueryDto query)
         {
             Expression<Func<Post, bool>> filter = x => !x.IsDeleted;
-            var userId = _claimService.GetUserId();
-            if (!string.IsNullOrEmpty(queryDto.SearchText))
+            var userId = query.UserId;
+            if (string.IsNullOrEmpty(userId))
+                userId = _claimService.GetUserId();
+            var user = await _userRepository.GetFirstAsync(p => p.Id == userId);
+            if (user == null)
             {
-                filter = filter.And(x => x.Description.Contains(queryDto.SearchText));
+                throw new NotFoundException(ValidationTexts.NotFound.Format("User", userId));
             }
-            if (!string.IsNullOrEmpty(queryDto.UserId))
-            {
-                var user = await _userRepository.GetFirstAsync(p => p.Id == queryDto.UserId);
-                if (user == null)
-                {
-                    throw new NotFoundException(ValidationTexts.NotFound.Format("User", queryDto.UserId));
-                }
-                filter = filter.And(x => x.UserId == queryDto.UserId);
-            }
-            else
-                filter = filter.And(x => x.UserId == userId);
+
             var paginationResponse = await _postRepository.GetAllAsync(filter, p => p.OrderByDescending(p => p.CreatedOn)
-                                                            , queryDto.PageIndex
-                                                            , queryDto.PageSize,
+                                                            , query.PageIndex
+                                                            , query.PageSize,
                                                             p => p.IgnoreAutoIncludes().Include(c => c.PostMedias)
                                                             );
+            foreach (var post in paginationResponse.Items)
+            {
+                post.PostMedias = post.PostMedias.OrderBy(m => m.MediaOrder).ToList();
+            }
             return _mapper.Map<List<PostDto>>(paginationResponse.Items);
         }
         public async Task<PostDto> GetPostByIdAsync(Guid postId)
         {
             var post = await _postRepository.GetFirstOrDefaultAsync(p => p.Id == postId && !p.IsDeleted
-                                                                    , p => p.Include(c => c.PostMedias).Include(c => c.Comments));
+                                                                    , p => p.Include(c => c.PostMedias).Include(c => c.Comments).Include(c => c.User));
             if (post == null)
                 throw new NotFoundException(ValidationTexts.NotFound.Format("Post", postId));
-            post.Comments = post.Comments.Where(p => p.CommentType == CommentType.Parent).ToList();
+            post.Comments = post.Comments.Where(p => p.CommentType == CommentType.Parent.ToString()).ToList();
+            post.PostMedias = post.PostMedias.OrderBy(m => m.MediaOrder).ToList();
             return _mapper.Map<PostDto>(post);
         }
-        public Task<PostDto> GetReccomendationPostsAsync(PaginationFilter query)
+        public async Task<List<PostDto>> GetReccomendationPostsAsync(PostSearchQueryDto query)
         {
-            throw new NotImplementedException();
+            var userId = _claimService.GetUserId();
+            if (!string.IsNullOrEmpty(query.SearchText))
+            {
+                if (query.IsTag)
+                {
+                    var postSearchTag = await _postRepository.GetAllAsync(p => p.PostDetailTags.Any(p => p.Tag.Contains(query.SearchText)) && !p.IsDeleted && p.UserId != userId
+                                                , p => p.OrderByDescending(p => p.CreatedOn)
+                                                , query.PageIndex
+                                                , query.PageSize
+                                                , p => p.Include(p => p.PostDetailTags).Include(p => p.User).Include(p => p.PostMedias));
+                    foreach (var post in postSearchTag.Items)
+                    {
+                        post.PostMedias = post.PostMedias.OrderBy(m => m.MediaOrder).ToList();
+                    }
+                    return _mapper.Map<List<PostDto>>(postSearchTag.Items);
+                }
+                var postSearchDescription = await _postRepository.GetAllAsync(p => p.Description.Contains(query.SearchText) && !p.IsDeleted && p.UserId != userId
+                                                , p => p.OrderByDescending(p => p.CreatedOn)
+                                                , query.PageIndex
+                                                , query.PageSize
+                                                , p => p.Include(p => p.PostDetailTags).Include(p => p.User).Include(p => p.PostMedias));
+                foreach (var post in postSearchDescription.Items)
+                {
+                    post.PostMedias = post.PostMedias.OrderBy(m => m.MediaOrder).ToList();
+                }
+                return _mapper.Map<List<PostDto>>(postSearchDescription.Items);
+            }
+
+            var posts = await _postRepository.GetAllAsync(p => p.UserId != userId && !p.IsDeleted, p => p.OrderByDescending(p => p.CreatedOn)
+                                                , query.PageIndex
+                                                , query.PageSize
+                                                , p => p.Include(p => p.PostDetailTags).Include(p => p.User).Include(p => p.PostMedias));
+            foreach (var post in posts.Items)
+            {
+                post.PostMedias = post.PostMedias.OrderBy(m => m.MediaOrder).ToList();
+            }
+            return _mapper.Map<List<PostDto>>(posts.Items);
         }
         public async Task<List<CommentDto>> GetAllReplyCommentsAsync(Guid postId, Guid commentId)
         {
-            var commentParent = _commentRepository.GetFirstOrDefaultAsync(p => p.Id == commentId && p.CommentType == CommentType.Parent);
+            var commentParent = _commentRepository.GetFirstOrDefaultAsync(p => p.Id == commentId && p.CommentType == CommentType.Parent.ToString());
             if (commentParent == null)
                 throw new NotFoundException(ValidationTexts.NotFound.Format("Comment", commentId));
             var comments = await _commentRepository.GetAllAsync(p => p.PostId == postId && p.ParentCommentId == commentId);
@@ -112,14 +143,17 @@ namespace ChitChat.Application.Services
             var requestPostMedia = await _cloudinaryService.PostMediaToCloudAsync(requestDto.Files, post.Id);
             var postMedias = _mapper.Map<List<PostMedia>>(requestPostMedia);
             await _postRepository.AddAsync(post);
+            int i = 0;
             foreach (var postMedia in postMedias)
             {
                 postMedia.PostId = post.Id;
+                postMedia.MediaOrder = i++;
             }
+
             await _postMediaRepository.AddRangeAsync(postMedias);
             return _mapper.Map<PostDto>(post);
         }
-        public async Task<CommentDto> CreateNewCommentAsync(Guid postId, CreateCommentRequestDto requestDto)
+        public async Task<CommentDto> CreateNewCommentAsync(Guid postId, CommentRequestDto requestDto)
         {
             var userId = _claimService.GetUserId();
             var comment = _mapper.Map<Comment>(requestDto);
@@ -134,20 +168,20 @@ namespace ChitChat.Application.Services
             };
             comment.PostId = postId;
             comment.UserPostedId = userId;
-            comment.CommentType = CommentType.Parent;
+            comment.CommentType = CommentType.Parent.ToString();
             ++post.CommentCount;
             await _commentRepository.AddAsync(comment);
             await _postRepository.UpdateAsync(post);
             await _userInteractionRepository.AddAsync(userInteraction);
             return _mapper.Map<CommentDto>(comment);
         }
-        public async Task<CommentDto> CreateReplyCommentAsync(Guid postId, Guid parentCommentId, CreateCommentRequestDto requestDto)
+        public async Task<CommentDto> CreateReplyCommentAsync(Guid postId, Guid parentCommentId, CommentRequestDto requestDto)
         {
-            if (!(await _commentRepository.AnyAsync(p => p.Id == parentCommentId && p.CommentType == CommentType.Parent)))
+            if (!(await _commentRepository.AnyAsync(p => p.Id == parentCommentId && p.CommentType == CommentType.Parent.ToString())))
                 throw new NotFoundException(ValidationTexts.NotFound.Format("Comment", parentCommentId));
             var userId = _claimService.GetUserId();
             var comment = _mapper.Map<Comment>(requestDto);
-            comment.CommentType = CommentType.Reply;
+            comment.CommentType = CommentType.Reply.ToString();
             var post = await _postRepository.GetFirstAsync(p => p.Id == postId);
             if (post == null)
                 throw new NotFoundException(ValidationTexts.NotFound.Format("Post", postId));
@@ -168,13 +202,13 @@ namespace ChitChat.Application.Services
         }
         #endregion
         #region PUT
-        public async Task<PostDto> UpdatePostByIdAsync(PostDto postDto, Guid postId)
+        public async Task<PostDto> UpdatePostByIdAsync(UpdatePostRequestDto postDto, Guid postId)
         {
             Post post = await _postRepository.GetFirstAsync(p => p.Id == postId);
-            if (post.Id != postId)
-                throw new InvalidModelException(ValidationTexts.NotValidate.Format("Post", postId));
             if (post == null)
                 throw new NotFoundException(ValidationTexts.NotFound.Format("Post", postId));
+            if (post.UserId != _claimService.GetUserId())
+                throw new ForbiddenException(ValidationTexts.Forbidden.Format("Post", postId));
             post = _mapper.Map(postDto, post);
             await _postRepository.UpdateAsync(post);
             return _mapper.Map<PostDto>(post);
@@ -221,11 +255,13 @@ namespace ChitChat.Application.Services
                 return false;
             }
         }
-        public async Task<CommentDto> UpdateCommentAsync(CommentDto commentDto, Guid commentId)
+        public async Task<CommentDto> UpdateCommentAsync(CommentRequestDto commentDto, Guid commentId)
         {
             var comment = await _commentRepository.GetFirstOrDefaultAsync(p => p.Id == commentId);
             if (comment == null)
                 throw new NotFoundException(ValidationTexts.NotFound.Format("Comment", commentId));
+            if (comment.UserPostedId != _claimService.GetUserId())
+                throw new ForbiddenException(ValidationTexts.Forbidden.Format("Comment", commentId));
             comment = _mapper.Map(commentDto, comment);
             await _commentRepository.UpdateAsync(comment);
             return _mapper.Map<CommentDto>(comment);
@@ -267,12 +303,15 @@ namespace ChitChat.Application.Services
             Post post = await _postRepository.GetFirstAsync(p => p.Id == postId);
             if (post == null)
                 throw new NotFoundException(ValidationTexts.NotFound.Format("Post", postId));
+            if (post.UserId != _claimService.GetUserId())
+                throw new ForbiddenException(ValidationTexts.Forbidden.Format("Post", postId));
             var postMedia = await _postMediaRepository.GetAllAsync(p => p.PostId == postId);
             if (postMedia != null)
             {
                 foreach (var media in postMedia)
                 {
                     await _postMediaRepository.DeleteAsync(media);
+                    await _cloudinaryService.DeleteMediaFromCloudAsync(postId, media.Id);
                 }
             }
             await _postRepository.DeleteAsync(post);
@@ -283,6 +322,8 @@ namespace ChitChat.Application.Services
             var comment = await _commentRepository.GetFirstOrDefaultAsync(p => p.Id == commentId);
             if (comment == null)
                 throw new NotFoundException(ValidationTexts.NotFound.Format("Comment", commentId));
+            if (comment.UserPostedId != _claimService.GetUserId())
+                throw new ForbiddenException(ValidationTexts.Forbidden.Format("Comment", commentId));
             await _commentRepository.DeleteAsync(comment);
             return _mapper.Map<CommentDto>(comment);
         }
